Update version reading from remote source and upgrade dependencies
This commit changes the way the version manifest is read. Instead of reading from a static local file, it is now read from a remote source to reliably get the latest data.

Additional changes in this commit are focused on compatibility and maintenance:
- Gradle version was upgraded from 8.3 to 8.5
- Java version was upgraded from 20 to 21
- The group name and version in build.gradle were updated
- Some dependency versions were also updated in the file.

Additionally, some IntelliJ IDEA settings files were updated and Git was set as the VCS.